+++ scripts/minify-dist.mjs
@@ -63,8 +63,7 @@ for(const file of [ 'app.css.dat', 'app.js.dat' ])
 
 const publicLoaderAssets = [
     [ 'src/assets/images/loading/octane-logo.png', 'octane-logo.png' ],
-    [ 'src/assets/images/loading/loading.gif', 'loading.gif' ],
-    [ 'src/assets/images/notifications/nitro_v3.png', 'nitro_v3.png' ]
+    [ 'src/assets/images/loading/loading.gif', 'loading.gif' ]
 ];
 
 for(const [ source, file ] of publicLoaderAssets)